display: bring back interpolated doze brightness for QHD panel
This was dropped in 22a180d5df0d955288a02695f6b4b5819eabdf85.

We have had custom low power AOD modes working on qhd panel variants,
however we hit a touchpanel firmware bug in these modes that we are
unlikely to workaround.

Due to how unlikely it is we will get proper AOD modes on this panel,
cleanup and bring back a diff impl of interpolated doze brightness for
this panel.

This is reimplemented with the following changes:

 - utilize aod_state
 - block backlight updates during lp*
 - use current aod brightness for fod dim
 - move aod_state to a u32 which tracks aod brightness on qhd panel:
   - this stops us from setting a too low doze brightness going
     doze->prox panel off->doze